#16 #17 #18 - Refactored some code. Moved betting routes that used to be part of users into bets.
