+++ public/javascript/play.js
@@ -6,24 +6,21 @@ var placeBet = () => {
 }
 
 var getBetValues = () => {
-        var bet = {
-            "userId": "",
-            "roundId": 1,
-            "games": []
-        };
-        for (i = 0; i < 6; i++) {
-            let teamATries = document.querySelector(`[game_block_${i}] input[team_A_tries]`)
-            let teamBTries = document.querySelector(`[game_block_${i}] input[team_B_tries]`)
-            let gameVictor = document.querySelector(`[game_block_${i}] select[game_1_winner]`)
-            bet.games.push({
-                teamATries: teamATries.value,
-                teamBTries: teamBTries.value,
-                gameVictor: gameVictor.value
-            });
-        };
-        let goldenTryInput = document.querySelector('input[golden_try_input]');
+    var bet = {
+        "games": [],
+        "goldenTrySelection": null
+    };
+    for (i = 0; i < 6; i++) {
+        let teamATries = document.querySelector(`[game_block_${i}] input[team_A_tries]`)
+        let teamBTries = document.querySelector(`[game_block_${i}] input[team_B_tries]`)
+        let gameVictor = document.querySelector(`[game_block_${i}] select[game_winner]`)
         bet.games.push({
-                goldenTrySelection: goldenTryInput.value
+            teamATries: teamATries.value,
+            teamBTries: teamBTries.value,
+            gameVictor: gameVictor.value
         });
-            return bet;
-        };
+    };
+    let goldenTryInput = document.querySelector('input[golden_try_input]');
+    bet.goldenTrySelection = goldenTryInput.value;
+    return bet;
+};
