+++ routes/users.js
@@ -1,13 +1,14 @@
 const express = require('express');
 const userService = require('../services/users');
 const cookieParser = require('cookie-parser');
-const gameService = require('../services/game');
-const roundsService = require('../services/rounds');
 
 const router = express.Router();
 
 router.get('/', function (req, res, next) {
-    // List all users - NO AUTH AT PRESENT
+    // List all users - NO AUTH AT PRESENT (only for admin)
+
+
+    // TODO: move the code below to a service (much like service/games.js and such)
     const db = req.app.get('super6db');
     const collection = db.collection('users');
     collection.find({}).toArray(function (err, docs) {
@@ -18,10 +19,27 @@ router.get('/', function (req, res, next) {
     });
 });
 
+router.get('/profile', function (req, res, next) {
+    // Should show the user profile page
+
+    // TODO: move the code below to a service (much like service/games.js and such)
+    // const db = req.app.get('super6db');
+    // const collection = db.collection('users');
+    // collection.find({}).toArray(function (err, docs) {
+    //     if (err) {
+    //         console.error(err);
+    //     }
+    //     res.json(docs)
+    // });
+    res.render('profile');
+});
+
 router.post('/signup', function (req, res, next) {
     // Save a user - NO AUTH AT PRESENT
     const email = req.body.email;
     const password = req.body.password;
+
+    // TODO: move the code below to a service (much like service/games.js and such)
     const db = req.app.get('super6db');
     userService.createUser(db, email, password, false, function (error, result) {
         res.cookie('super6token', 'abcd1234', { maxAge: 3600000 });
@@ -36,6 +54,8 @@ router.post('/signup', function (req, res, next) {
 router.post('/login/', function (req, res, next) {
     const email = req.body.email;
     const password = req.body.password;
+
+    // needs to vberify the users in the DB (do it via a service)
     if (email === 'user@gmail.com' && password === 'password') { // to be replaced with call to the user service
         res.cookie('super6token', 'abcd1234', { maxAge: 3600000 }); // to be replaced with the session token generated by the service
         res.redirect('../');
@@ -50,25 +70,5 @@ router.get('/logout', function (req, res, next) {
     res.redirect('../');
 });
 
-router.get('/play', function (req, res, next) {
-    const db = req.app.get('super6db');
-    gameService.fetchFuture(db).then((games) => {
-        res.render('play', {
-            title: 'Super6 Rugby - Play', games: games, testValue: "HELLO WORLD", loggedIn: true
-        }); // TODO: Logged in needs to reflect cookie value and checked against db and games need to be pushed
-    },
-        (reason) => {
-            console.log(reason)
-        });
-});
-
-router.get('/history', function (req, res, next) {
-    const db = req.app.get('super6db');
-    roundsService.fetch(db).then((rounds) => {
-        gameService.fetchByRound(db, rounds).then((games) => {
-            res.render('history', { title: 'Super6 Rugby - Your History', loggedIn: true, rounds: rounds, games: games });
-        });
-    });
-});
 
 module.exports = router;
