+++ views/play.ejs
@@ -20,15 +20,18 @@
         </div>
 
         <!-- <% include common/games.ejs %> -->
-        <% for (i = 0; i < 6; i++) { %>
+        <% for (let i = 0; i < games.length; i++) { %>
         <div game_block game_block_<%=i%> class="game_row">
             <div class="column">
                 <h2>TEAMA VS TEAMB</h2>
             </div>
             <div class="column">
-                Team A tries: <input team_A_tries name="bock<%=i%>TeamATries" type="number" min="0" size="4">
-                Team B tries: <input team_B_tries name="block<%=i%>TeamBTries" type="number" min="0" size="4"><br>
-                Select Match Winner:<select game_1_winner name="Winner">
+                <%= games[i].teamA.name %> tries:
+                <input team_A_tries name="block<%=i%>TeamATries" type="number" min="0" size="4">
+                <%= games[i].teamB.name %> tries:
+                <input team_B_tries name="block<%=i%>TeamBTries" type="number" min="0" size="4">
+                <br>
+                Select Match Winner:<select game_winner name="Winner">
                     <option value="TEAM A">TEAM A</option>
                     <option value="TEAM B">TEAM B</option>
                 </select>
@@ -37,14 +40,22 @@
         <% } %>
     </div>
 
+    <% if (games.length > 0) { %>
     <div class="golden_try" class="game_row">
         <h2>Golden Try Prediction</h2> <input golden_try_input name="goldenTry" type="number" min="0" size="5">
     </div>
+    <% } %>
 
     <div class='place_bet_button'>
-        <button onclick="placeBet()" type="button" id="bet_button">
+        <% if (games.length > 0) { %>
+        <button onclick="placeBet(<%= games[0].roundId %>)" type="button" id="bet_button">
             <h2>Place your Bets!</h2>
         </button>
+        <% } else { %>
+        <button disabled type="button" id="bet_button">
+            <h2>No games!</h2>
+        </button>
+        <% } %>
     </div>
 
     <%- include common/footer.ejs %>
