+++ views/profile.ejs
@@ -1,0 +1,6 @@
+TODO!<br />
+<br />
+Include user information<br />
+This probably should look like a dashboard.<br />
+<br />
+Have a link to <a href="/bets/history">/bets/history</a>?
